Attempt to build GTK3 with travis CI. Refs #6.
The GTK3 variant of this library is not on hackage and is not
being checked or pushed by the Travis CI. This makes it unavailable
in practice for people trying to use Keera Hails.

This commit adds a new variable to the build matrix to compile both
for GTK2 and GTK3, and to deploy both.

It also adds an upper bound and a description to the cabal file
to make cabal check pass.