Update Travis to build any develop* using Haskell. Refs #6.
Travis is refusing to build branches with names other than master and develop
because of the configuration setup. This makes it impossible to know if a build
compiles on the CI without merging to the main branch.

Also, the setup is set to use C, which means some steps must be manually
done, and only compiles for GTK2.

This commit alters the Travis CI configuration to:
- Build any branch with the name develop.*, where . represents any character.
- Use Haskell as the Travis image, making it less cumbersome to install.
- Install missing dependencies for GTK3.
- Disable support of GHC 7.6, 7.8 and 8.0, and enable support for GHC 8.2 and
  8.4 (see related bug https://github.com/keera-studios/keera-hails/issues/25).